remove mkl baxed CSR creation

+++ cpp/knng/math_operations.hpp
@@ -82,14 +82,14 @@ class MathOp {
 #pragma omp parallel for
     for (INDEX_TYPE i = 0; i < rows; i++) {
       for (INDEX_TYPE j = 0; j < cols; j++) {
-        uint64_t access_index = static_cast<uint64_t>(j + i * cols);
+        uint64_t access_index = static_cast<uint64_t>(j) + static_cast<uint64_t>(i) * static_cast<uint64_t>(cols);
         arr[access_index] = 0.0;
       }
     }
 #pragma omp parallel for
     for (INDEX_TYPE i = 0; i < rows; i++) {
       for (INDEX_TYPE j = 0; j < cols; j++) {
-        uint64_t access_index = static_cast<uint64_t>(j + i * cols);
+        uint64_t access_index = static_cast<uint64_t>(j) + static_cast<uint64_t>(i) * static_cast<uint64_t>(cols);
         arr[access_index] = (*data)[j][i];
       }
     }
